pyproject.toml
[tool.poetry]
name = "backend-ad-task"
version = "0.1.0"
description = ""
authors = ["Alteian <Alteian@proton.me>"]
readme = "README.md"
packages = [
    { include = "*", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.12"
granian = "^1.2.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.6.4"
python-decouple = "^3.8"
asyncio = "^3.4.3"
redis = "^5.0.3"
celery = "^5.3.6"


[tool.poetry.group.dev.dependencies]
granian = {extras = ["reload"], version = "^1.2.0"}
mypy = "^1.8.0"
ruff = "^0.1.14"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff.format]

[tool.ruff.isort]
known-third-party = []
extra-standard-library =  ["typing_extensions"]

[tool.ruff]
line-length = 120
target-version = "py312"
select = ["ALL"]
ignore = [
    # https://github.com/astral-sh/ruff/pull/4427
    # equivalent to keep-runtime-typing
    "UP006",
    "UP007",

    "TID252",
    # we use asserts in tests and to hint mypy
    "S101",
    "S102",
    "S104",
    "S324",
    # maybe we can enable this in future
    # we'd want to have consistent docstrings in future
    "D",
    "ANN101", # missing annotation for self?
    # definitely enable these, maybe not in tests
    "ANN102",
    "ANN202",
    "ANN204",
    "ANN205",
    "ANN401",
    "PGH003",
    "PGH004",
    "RET504",
    "RET505",
    "RET506",
    "RET507",
    "RET503",
    "BLE001",
    "B008",
    "N811",
    "N804",
    "N818",
    # Variable `T` in function should be lowercase
    # this seems a potential bug or opportunity for improvement in ruff
    "N806",

    # first argument should named self (found in tests)
    "N805",

    "N815",

    # shadowing builtins
    "A001",
    "A002",
    "A003",

    "ARG001",
    "ARG002",
    "ARG003",
    "ARG004",
    "ARG005",
    "FBT001",
    "FBT002",
    "FBT003",

    "PT001",
    "PT023",

    # this is pretty much handled by black
    "E501",

    # enable these, we have some in tests
    "B006",
    "PT004",
    "PT007",
    "PT011",
    "PT012",
    "PT015",
    "PT017",
    "C414",
    "N802",

    "SIM117",
    "SIM102",

    "F841",
    "B027",
    "B905",
    "ISC001",

    # same?
    "S105",
    "S106",

    "DTZ003",
    "DTZ005",

    "RSE102",
    "SLF001",

    # in tests
    "DTZ001",

    "EM101",
    "EM102",
    "EM103",

    "B904",
    "B019",

    "N801",
    "N807",

    # pandas
    "PD",

    "RUF012",
    "PLC0105",
    "FA102",

    # code complexity
    "C",
    "C901",

    # trailing commas
    "COM812",

    "PLR",
    "INP",
    "TRY",
    "SIM300",
    "SIM114",

    "DJ008",
    "TD002",
    "TD003",
    "FIX001",
    "FIX002",
    "FA100",
]
fix = true
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
src = ["src"]


[tool.mypy]
python_version = "3.12"
files = "src"
plugins = ["pydantic.mypy",]
implicit_reexport = false
warn_unused_configs = true
warn_unused_ignores = true
check_untyped_defs = true
ignore_errors = false
strict_optional = true
show_error_codes = true
warn_redundant_casts = true
ignore_missing_imports = true
install_types = true
non_interactive = true
show_traceback = true
enable_incomplete_feature = "Unpack"
